refactor!: remove QueryVisualize components
