--- frontend/src/components/Query/Visualize/NumberCard.vue
@@ -1,43 +1,0 @@
-<script setup>
-import { computed, inject } from 'vue'
-const props = defineProps({
-	data: {
-		type: Object,
-		required: true,
-	},
-	options: {
-		type: Object,
-		default: {},
-	},
-})
-const $utils = inject('$utils')
-
-const formattedValue = computed(() => {
-	let value = props.data.value || 0
-	value = $utils.getShortNumber(value, props.options.decimals)
-	return value
-})
-
-const prefix = computed(() => {
-	return props.options.prefix || ''
-})
-const suffix = computed(() => {
-	return props.options.suffix || ''
-})
-</script>
-
-<template>
-	<div class="h-full w-full overflow-hidden rounded-md border py-5 px-8">
-		<div class="mx-auto flex h-full w-full min-w-40 flex-col justify-center overflow-hidden">
-			<div
-				:title="props.data.title"
-				class="w-full overflow-hidden text-ellipsis whitespace-nowrap text-base text-gray-600"
-			>
-				{{ props.data.title }}
-			</div>
-			<div class="text-[38px] leading-tight">
-				{{ prefix }}{{ formattedValue }}{{ suffix }}
-			</div>
-		</div>
-	</div>
-</template>
